ADK-STREAMING/requirements.txt
google-adk==1.25.1
google-genai>=1.65.0
python-dotenv
uvicorn[standard]
